+++ src/ArduinoJson/Operators/VariantOr.hpp
@@ -15,8 +15,7 @@ class VariantOr {
  public:
   // Returns the default value if the VariantRef is undefined of incompatible
   template <typename T>
-  typename enable_if<!is_integral<T>::value, T>::type operator|(
-      const T &defaultValue) const {
+  T operator|(const T &defaultValue) const {
     if (impl()->template is<T>())
       return impl()->template as<T>();
     else
@@ -28,17 +27,6 @@ class VariantOr {
   const char *operator|(const char *defaultValue) const {
     const char *value = impl()->template as<const char *>();
     return value ? value : defaultValue;
-  }
-
-  // Returns the default value if the VariantRef is undefined of incompatible
-  // Special case for integers: we also accept double
-  template <typename Integer>
-  typename enable_if<is_integral<Integer>::value, Integer>::type operator|(
-      const Integer &defaultValue) const {
-    if (impl()->template is<double>())
-      return impl()->template as<Integer>();
-    else
-      return defaultValue;
   }
 
  private: