+++ test/JsonVariant/or.cpp
@@ -51,6 +51,12 @@ TEST_CASE("JsonVariant::operator|()") {
     REQUIRE(result == "default");
   }
 
+  SECTION("int | uint8_t (out of range)") {
+    variant.set(666);
+    uint8_t result = variant | static_cast<uint8_t>(42);
+    REQUIRE(result == 42);
+  }
+
   SECTION("int | int") {
     variant.set(0);
     int result = variant | 666;
@@ -58,8 +64,9 @@ TEST_CASE("JsonVariant::operator|()") {
   }
 
   SECTION("double | int") {
-    variant.set(42.0);
-    int result = variant | 666;
+    // NOTE: changed the behavior to fix #981
+    variant.set(666.0);
+    int result = variant | 42;
     REQUIRE(result == 42);
   }